add stats.h for bitstream, and fix jenkins warnings

+++ Framework/DataHandling/inc/MantidDataHandling/LoadDNSEvent.h
@@ -83,7 +83,7 @@ private:
   };
 
   struct NeutronEvent {
-    uint8_t wsIndex = 0;
+    size_t wsIndex = 0;
     CompactEvent event = {};
   };
